+++ controllers/clustervectorpipeline_controller.go
@@ -19,18 +19,24 @@ package controllers
 import (
 	"context"
 
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/kubernetes"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	observabilityv1alpha1 "github.com/kaasops/vector-operator/api/v1alpha1"
+	"github.com/go-logr/logr"
+	vectorv1alpha1 "github.com/kaasops/vector-operator/api/v1alpha1"
 )
 
 // ClusterVectorPipelineReconciler reconciles a ClusterVectorPipeline object
 type ClusterVectorPipelineReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
+
+	// Temp. Wait this issue - https://github.com/kubernetes-sigs/controller-runtime/issues/452
+	Clientset *kubernetes.Clientset
 }
 
 //+kubebuilder:rbac:groups=observability.kaasops.io,resources=clustervectorpipelines,verbs=get;list;watch;create;update;patch;delete
@@ -47,16 +53,74 @@ type ClusterVectorPipelineReconciler struct {
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.12.2/pkg/reconcile
 func (r *ClusterVectorPipelineReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	_ = log.FromContext(ctx)
+	log := log.FromContext(ctx).WithValues("VectorPipeline", req.Name)
 
-	// TODO(user): your logic here
+	log.Info("start Reconcile ClusterVectorPipeline")
 
+	// cvp, done, result, err := r.findClusterVectorPipelineCustomResourceInstance(ctx, log, req)
+	// if done {
+	// 	return result, err
+	// }
+	// hash, err := vectorpipeline.GetSpecHash(cvp.Spec)
+	// if err != nil {
+	// 	return ctrl.Result{}, err
+	// }
+	// if cvp.Status.LastAppliedPipelineHash != nil && *hash == *cvp.Status.LastAppliedPipelineHash {
+	// 	log.Info("ClusterVectorPipeline has no changes. Finish Reconcile ClusterVectorPipeline")
+	// 	return ctrl.Result{}, nil
+	// }
+
+	// vectorInstances := &vectorv1alpha1.VectorList{}
+	// err = r.List(ctx, vectorInstances)
+	// if err != nil {
+	// 	return ctrl.Result{}, err
+	// }
+
+	// if len(vectorInstances.Items) == 0 {
+	// 	log.Info("Vertors not found")
+	// 	return ctrl.Result{}, nil
+	// }
+
+	// for _, v := range vectorInstances.Items {
+	// 	if v.DeletionTimestamp != nil {
+	// 		continue
+	// 	}
+	// 	if err = checkConfig(ctx, &v, vp, r.Client, r.Clientset); err != nil {
+	// 		return ctrl.Result{}, err
+	// 	}
+	// 	if err = vectorpipeline.SetLastAppliedPipelineStatus(ctx, vp, r.Client); err != nil {
+	// 		return ctrl.Result{}, err
+	// 	}
+
+	// }
+
+	log.Info("finish Reconcile ClusterVectorPipeline")
 	return ctrl.Result{}, nil
+}
+
+func (r *ClusterVectorPipelineReconciler) findClusterVectorPipelineCustomResourceInstance(ctx context.Context, log logr.Logger, req ctrl.Request) (*vectorv1alpha1.ClusterVectorPipeline, bool, ctrl.Result, error) {
+	// fetch the master instance
+	cvp := &vectorv1alpha1.ClusterVectorPipeline{}
+	err := r.Get(ctx, req.NamespacedName, cvp)
+	if err != nil {
+		if errors.IsNotFound(err) {
+			// Request object not found, could have been deleted after reconcile request.
+			// Owned objects are automatically garbage collected. For additional cleanup logic use finalizers.
+			// Return and don't requeue
+			log.Info("ClusterVectorPipeline CR not found. Ignoring since object must be deleted")
+			return nil, true, ctrl.Result{}, nil
+		}
+		// Error reading the object - requeue the request.
+		log.Error(err, "Failed to get Vector")
+		return nil, true, ctrl.Result{}, err
+	}
+	log.Info("Get Vector Pipeline" + cvp.Name)
+	return cvp, false, ctrl.Result{}, nil
 }
 
 // SetupWithManager sets up the controller with the Manager.
 func (r *ClusterVectorPipelineReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&observabilityv1alpha1.ClusterVectorPipeline{}).
+		For(&vectorv1alpha1.ClusterVectorPipeline{}).
 		Complete(r)
 }